warnings.

+++ test/aiger.cpp
@@ -43,16 +43,16 @@ namespace benchmarks
   struct aig_info
   {
     /* header information */
-    uint32_t maximum_variable_index{0};
-    uint32_t number_of_inputs{0};
-    uint32_t number_of_latches{0};
-    uint32_t number_of_outputs{0};
-    uint32_t number_of_ands{0};
+    uint64_t maximum_variable_index{0};
+    uint64_t number_of_inputs{0};
+    uint64_t number_of_latches{0};
+    uint64_t number_of_outputs{0};
+    uint64_t number_of_ands{0};
     
-    uint32_t number_of_bad_states{0};
-    uint32_t number_of_constraints{0};
-    uint32_t number_of_justice{0};
-    uint32_t number_of_fairness{0};
+    uint64_t number_of_bad_states{0};
+    uint64_t number_of_constraints{0};
+    uint64_t number_of_justice{0};
+    uint64_t number_of_fairness{0};
 
     /* count calls */
     uint32_t on_input_calls{0};
